Adds omit to common/utils
Adds omit to common as an alternative to lodash.omit.
